add jdk-test target, and fix failures
The intent of this target is to run our test suite against the installed jre.
This should help prevent our VM from diverging in implementation from the jdk.

The remainder of this commit fixes the problems that this exposes.

+++ classpath/java/util/zip/ZipFile.java
@@ -298,7 +298,7 @@ public class ZipFile {
       }
     }
 
-    public int getCompressedSize() {
+    public long getCompressedSize() {
       try {
         return compressedSize(window, pointer);
       } catch (IOException e) {
@@ -306,7 +306,7 @@ public class ZipFile {
       }
     }
 
-    public int getSize() {
+    public long getSize() {
       try {
         return uncompressedSize(window, pointer);
       } catch (IOException e) {
